#44 move utils func to separated package
