+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/ios/IosMRGenerator.kt
@@ -57,7 +57,7 @@ class IosMRGenerator(
 
     override fun getImports(): List<ClassName> = listOf(
         bundleClassName,
-        ClassName(packageName = "dev.icerock.moko.resources", simpleName = "loadableBundle")
+        ClassName(packageName = "dev.icerock.moko.resources.utils", simpleName = "loadableBundle")
     )
 
     override fun apply(generationTask: Task, project: Project) {
