+++ resources/src/iosMain/kotlin/dev/icerock/moko/resources/utils/NSBundleExt.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.resources
+package dev.icerock.moko.resources.utils
 
 import platform.Foundation.NSBundle
